requirements.txt
# Python dependencies for HP cartridge classifier
tensorflow
numpy
scikit-learn
matplotlib
seaborn
pillow
